backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.134.0",
    "motor>=3.6.1",
    "dnspython>=2.7.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]
